oops

+++ Task2/MergeSort.h
@@ -12,6 +12,6 @@
 
 using namespace std;
 
-void MergeSort(int *arr, int size, int left, int right);
+void MergeSort(int *arr, long long size, int left, int right);
 
 #endif //FINALTEST_MERGESORT_H
